Add optional query parameter to Connection#table
Allows user to specify a query for the table. For example, they can
filter or specify an order for the rows returned. This query needs to be
formatted according to the PostgREST API standards. So a query of { id:
'eq.1' } will return records with an ID equal to 1; however, { id: 1 }
will return an error message.